# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dev/code/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工具模块"
                ],
                "summary": "生成代码",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd.CodeCreateCmd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/dev/code/download": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工具模块"
                ],
                "summary": "生成代码",
                "parameters": [
                    {
                        "description": "生成代码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd.CodeCreateCmd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/dev/code/page": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工具模块"
                ],
                "summary": "查询数据库表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字值",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/sys/menu/list": {
            "get": {
                "tags": [
                    "菜单模块"
                ],
                "summary": "分页",
                "parameters": [
                    {
                        "description": "菜单列表",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/cmd.MenuListQry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/sys/menu/query": {
            "get": {
                "tags": [
                    "菜单模块"
                ],
                "summary": "下拉框数据",
                "parameters": [
                    {
                        "description": "菜单列表",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/cmd.MenuListQry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd.CodeCreateCmd": {
            "type": "object",
            "properties": {
                "createType": {
                    "type": "integer"
                },
                "module": {
                    "type": "string"
                },
                "tableNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "cmd.MenuListQry": {
            "type": "object",
            "properties": {
                "parentUuid": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}